Fixed log10 error
Was having a problem where extracting each digit using the method where you figure out the current power was not dependable as with larger numbers the math.log10 function would round up. What I ended up doing is just converting the number to a string and then iterating through a it and appending it to a list. Apperently it is also faster like so, therefore it is a complete DUB